implement the AAA interest rate generator (#66)
* update docstring for NelsonSiegelInterpolator

* update StochasticProcess docstring

* create AcademyRateModel

This model implements the American Academy of Actuaries stochastic
interest rate model, using the existing AcademyRateProcess. The model
also handles the quirky interpolation and perturbation algorithms used
in the AAA generator.

* write functions for "interpolate" and perturb"

These functions are used internally by the AcademyRateModel class to do
post-processing of the raw scenarios from the AcademyRateProcess.

* minor improvements to interpolator doctstrings

Also updated some of the function signatures to consistently use capital
"X" in methods that require (or allow) arrays to be passed.

* add testing files for AcademyRateModel

* add a few unit tests

* add a sample scenario test to AcademyRateModel

This commit adds a new dataset - the academy_sample_scenario.csv - which
contains the values from a single scenario of the model. It also stores
the inputs used to generate the scenario, which is required to validate
the AcademyRateModel scenario generator.